Things aren't "moving," may need to adjust grid to a map with keys.
This version doesn't "move," in the way that the white dot ought to be "falling" from "gravity" at a rate of 200ms. Also, none of the flower/towers appear at all. This seems to be due to the slice method. Its seems that we need to keep track of the changes to each "cell" using Keys from the map maybe?